refactor: Moving named methods from the mapper to the mapper helper.
